Added dynamically changing bars based on reviews
