+++ src/ratings-and-reviews/components/Sort.jsx
@@ -1,19 +1,12 @@
 import React from 'react';
+import reviewsTotal from './ReviewBreakdown.jsx';
+
 // Number of reviews in label based on data from API
 // Depending on which option is selected, re sort the tiles in ReviewTile based on corresponding option
 
-/** @param {Object} reviews */
-const reviewsTotal = (reviews) => {
-  let total = 0;
-  for (let i in reviews) {
-    total += reviews[i];
-  }
-  return total;
-};
-
-const Sort = ({breakdown}) => (
+const Sort = ({total}) => (
   <div className="review-sort">
-    <label>{reviewsTotal(breakdown)} reviews, sorted by </label>
+    <label>{total} reviews, sorted by </label>
     <select className="review-dropdown">
       <option>Relevance</option>
       <option>Newest</option>
